Fix many bags: 1. Heap corruption because of prematurely deletion 2. Non virtual destructor in abstract class 3. Cache controll for suspended readers 4. Fix some comparison typos and forgoten checks

+++ HttpProxy/main.cpp
@@ -18,6 +18,6 @@ int main(int argc, char **argv)
 		return 1;
 	}
 	
-	HttpProxy proxy{ lport, threads };
+	HttpProxy proxy{ lport, static_cast<size_t>(threads) };
 	proxy.run();
 }
